+++ deltafi-ui/src/composables/useServerSentEvents.ts
@@ -19,21 +19,20 @@
 import ReconnectingEventSource from "reconnecting-eventsource";
 import { ref, Ref } from "vue";
 
+const connectionStatus = ref('CONNECTING') as Ref<'CONNECTING' | 'CONNECTED' | 'DISCONNECTED'>;
 const serverSentEvents = new ReconnectingEventSource("/api/v1/sse");
 
+serverSentEvents.addEventListener('open', () => {
+  connectionStatus.value = 'CONNECTED';
+});
+
+serverSentEvents.addEventListener('error', () => {
+  if (connectionStatus.value === 'CONNECTED') {
+    connectionStatus.value = 'DISCONNECTED';
+  }
+});
+
 export default function useServerSentEvents() {
-  const connectionStatus = ref('CONNECTING') as Ref<'CONNECTING' | 'CONNECTED' | 'DISCONNECTED'>;
-
-  serverSentEvents.addEventListener('open', () => {
-    connectionStatus.value = 'CONNECTED';
-  });
-
-  serverSentEvents.addEventListener('error', () => {
-    if (connectionStatus.value === 'CONNECTED') {
-      connectionStatus.value = 'DISCONNECTED';
-    }
-  });
-
   return {
     serverSentEvents,
     connectionStatus