+++ deltafi-ui/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.101.5]
+
+### Fixed
+
+- Remove `onCompletion` from snapshots
+
 ## [0.101.4]
 
 ### Changed